rename button link component file to link
