+++ src/renderer/src/routes/onboarding/project/create/$projectId.success.tsx
@@ -13,8 +13,8 @@ import {
 import { defineMessages, useIntl } from 'react-intl'
 
 import { GREEN, LIGHT_GREY } from '../../../../colors'
-import { ButtonLink } from '../../../../components/button-link'
 import { Icon } from '../../../../components/icon'
+import { ButtonLink } from '../../../../components/link'
 import { COMAPEO_CORE_REACT_ROOT_QUERY_KEY } from '../../../../lib/comapeo'
 
 export const Route = createFileRoute(
